extract cache-manager
